+++ aepsych/server/server.py
@@ -13,19 +13,28 @@ import sys
 import threading
 import traceback
 import warnings
-from collections.abc import Iterable
-
-import aepsych.database.db as db
-import aepsych.utils_logging as utils_logging
-
 import dill
 import numpy as np
 import pandas as pd
 import torch
+
+import aepsych.database.db as db
+import aepsych.utils_logging as utils_logging
+
 from aepsych.config import Config
 from aepsych.server.sockets import BAD_REQUEST, createSocket, DummySocket
 from aepsych.strategy import AEPsychStrategy, SequentialStrategy
 from aepsych.version import __version__
+
+from aepsych.server.message_handlers.handle_ask import handle_ask, ask, handle_ask_v01
+from aepsych.server.message_handlers.handle_setup import (
+    handle_setup,
+    handle_setup_v01,
+    handle_get_config,
+    configure,
+)
+from aepsych.server.message_handlers.handle_query import handle_query
+from aepsych.server.message_handlers.handle_tell import handle_tell
 
 logger = utils_logging.getLogger(logging.INFO)
 DEFAULT_DESC = "default description"
@@ -100,7 +109,7 @@ class AEPsychServer(object):
             self.socket.send(result)
         else:
             if self.can_pregen_ask and (len(self._pregen_asks) == 0):
-                self._pregen_asks.append(self.ask())
+                self._pregen_asks.append(ask(self))
 
     def serve(self):
         """Run the server. Note that all configuration outside of socket type and port
@@ -134,42 +143,6 @@ class AEPsychServer(object):
             self.cleanup()
             sys.exit(0)
 
-    def replay(self, uuid_to_replay, skip_computations=False):
-        """
-        Run a replay against the server. The UUID will be looked up in the database.
-        if skip_computations is true, skip all the asks and queries, which should make the replay much faster.
-        """
-        if uuid_to_replay is None:
-            raise RuntimeError("UUID is a required parameter to perform a replay")
-
-        if self.db is None:
-            raise RuntimeError("A database is required to perform a replay")
-
-        if skip_computations is True:
-            logger.info(
-                "skip_computations=True, make sure to refit the final strat before doing anything!"
-            )
-
-        master_record = self.db.get_master_record(uuid_to_replay)
-
-        if master_record is None:
-            raise RuntimeError(
-                f"The UUID {uuid_to_replay} isn't in the database. Unable to perform replay."
-            )
-
-        # this prevents writing back to the DB and creating a circular firing squad
-        self.is_performing_replay = True
-        self.skip_computations = skip_computations
-
-        for result in master_record.children_replay:
-            request = result.message_contents
-            logger.debug(f"replay - type = {result.message_type} request = {request}")
-            self.handle_request(request)
-
-        self._db_master_record = master_record
-        self.is_performing_replay = False
-        self.skip_computations = False
-
     def _unpack_strat_buffer(self, strat_buffer):
         if isinstance(strat_buffer, io.BytesIO):
             strat = torch.load(strat_buffer, pickle_module=dill)
@@ -184,132 +157,6 @@ class AEPsychServer(object):
         else:
             raise RuntimeError("Trying to load strat in unknown format!")
         return strat
-
-    def get_strats_from_replay(self, uuid_of_replay=None, force_replay=False):
-        if uuid_of_replay is None:
-            records = self.db.get_master_records()
-            if len(records) > 0:
-                uuid_of_replay = records[-1].experiment_id
-            else:
-                raise RuntimeError("Server has no experiment records!")
-
-        if force_replay:
-            warnings.warn(
-                "Force-replaying to get non-final strats is deprecated after the ability"
-                + " to save all strats was added, and will eventually be removed.",
-                DeprecationWarning,
-            )
-            self.replay(uuid_of_replay, skip_computations=True)
-            for strat in self._strats:
-                if strat.has_model:
-                    strat.model.fit(strat.x, strat.y)
-            return self._strats
-        else:
-            strat_buffers = self.db.get_strats_for(uuid_of_replay)
-            return [self._unpack_strat_buffer(sb) for sb in strat_buffers]
-
-    def get_strat_from_replay(self, uuid_of_replay=None, strat_id=-1):
-        if uuid_of_replay is None:
-            records = self.db.get_master_records()
-            if len(records) > 0:
-                uuid_of_replay = records[-1].experiment_id
-            else:
-                raise RuntimeError("Server has no experiment records!")
-
-        strat_buffer = self.db.get_strat_for(uuid_of_replay, strat_id)
-        if strat_buffer is not None:
-            return self._unpack_strat_buffer(strat_buffer)
-        else:
-            warnings.warn(
-                "No final strat found (likely due to old DB,"
-                + " trying to replay tells to generate a final strat. Note"
-                + " that this fallback will eventually be removed!",
-                DeprecationWarning,
-            )
-            # sometimes there's no final strat, e.g. if it's a very old database
-            # (we dump strats on crash) in this case, replay the setup and tells
-            self.replay(uuid_of_replay, skip_computations=True)
-            # then if the final strat is model-based, refit
-            strat = self._strats[strat_id]
-            if strat.has_model:
-                strat.model.fit(strat.x, strat.y)
-            return strat
-
-    def _flatten_tell_record(self, rec):
-        out = {}
-        out["response"] = int(rec.message_contents["message"]["outcome"])
-
-        out.update(
-            pd.json_normalize(
-                rec.message_contents["message"]["config"], sep="_"
-            ).to_dict(orient="records")[0]
-        )
-
-        if rec.extra_info is not None:
-            out.update(rec.extra_info)
-
-        return out
-
-    def get_dataframe_from_replay(self, uuid_of_replay=None, force_replay=False):
-        # DeprecationWarning
-        warnings.warn(
-            "get_dataframe_from_replay is deprecated."
-            + " Use generate_experiment_table with return_df = True instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-        if uuid_of_replay is None:
-            records = self.db.get_master_records()
-            if len(records) > 0:
-                uuid_of_replay = records[-1].experiment_id
-            else:
-                raise RuntimeError("Server has no experiment records!")
-
-        recs = self.db.get_replay_for(uuid_of_replay)
-
-        strats = self.get_strats_from_replay(uuid_of_replay, force_replay=force_replay)
-
-        out = pd.DataFrame(
-            [
-                self._flatten_tell_record(rec)
-                for rec in recs
-                if rec.message_type == "tell"
-            ]
-        )
-
-        # flatten any final nested lists
-        def _flatten(x):
-            return x[0] if len(x) == 1 else x
-
-        for col in out.columns:
-            if out[col].dtype == object:
-                out.loc[:, col] = out[col].apply(_flatten)
-
-        n_tell_records = len(out)
-        n_strat_datapoints = 0
-        post_means = []
-        post_vars = []
-
-        # collect posterior means and vars
-        for strat in strats:
-            if strat.has_model:
-                post_mean, post_var = strat.predict(strat.x)
-                n_tell_records = len(out)
-                n_strat_datapoints += len(post_mean)
-                post_means.extend(post_mean.detach().numpy())
-                post_vars.extend(post_var.detach().numpy())
-
-        if n_tell_records == n_strat_datapoints:
-            out["post_mean"] = post_means
-            out["post_var"] = post_vars
-        else:
-            logger.warn(
-                f"Number of tell records ({n_tell_records}) does not match "
-                + f"number of datapoints in strat ({n_strat_datapoints}) "
-                + "cowardly refusing to populate GP mean and var to dataframe!"
-            )
-        return out
 
     def generate_experiment_table(
         self, experiment_id, table_name="experiment_table", return_df=False
@@ -369,7 +216,7 @@ class AEPsychServer(object):
         handled_types = ["setup", "resume", "ask"]
         if request["type"] == "setup":
             if request["version"] == "0.01":
-                ret_val = self.handle_setup_v01(request)
+                ret_val = handle_setup_v01(self, request)
             else:
                 raise RuntimeError(
                     f"Unknown message version {request['version']} for message 'setup'!"
@@ -383,7 +230,7 @@ class AEPsychServer(object):
                 )
         elif request["type"] == "ask":
             if request["version"] == "0.01":
-                ret_val = self.handle_ask_v01(request)
+                ret_val = handle_ask_v01(self, request)
             else:
                 raise RuntimeError(
                     f"Unknown message version {request['version']} for message 'ask'!"
@@ -398,56 +245,6 @@ class AEPsychServer(object):
             return self.unversioned_handler(request)
         return ret_val
 
-    def handle_setup_v01(self, request):
-        logger.debug("got setup message!")
-        ### make a temporary config object to derive parameters because server handles config after table
-        if (
-            "config_str" in request["message"].keys()
-            or "config_dict" in request["message"].keys()
-        ):
-            tempconfig = Config(**request["message"])
-            if not self.is_performing_replay:
-                experiment_id = None
-                if self._db_master_record is not None:
-                    experiment_id = self._db_master_record.experiment_id
-                if "metadata" in tempconfig.keys():
-                    cdesc = (
-                        tempconfig["metadata"]["experiment_description"]
-                        if ("experiment_description" in tempconfig["metadata"].keys())
-                        else DEFAULT_DESC
-                    )
-                    cname = (
-                        tempconfig["metadata"]["experiment_name"]
-                        if ("experiment_name" in tempconfig["metadata"].keys())
-                        else DEFAULT_NAME
-                    )
-                    cid = (
-                        tempconfig["metadata"]["experiment_id"]
-                        if ("experiment_id" in tempconfig["metadata"].keys())
-                        else None
-                    )
-                    self._db_master_record = self.db.record_setup(
-                        description=cdesc,
-                        name=cname,
-                        request=request,
-                        id=cid,
-                        extra_metadata=tempconfig.jsonifyMetadata(),
-                    )
-                ### if the metadata does not exist, we are going to log nothing
-                else:
-                    self._db_master_record = self.db.record_setup(
-                        description=DEFAULT_DESC,
-                        name=DEFAULT_NAME,
-                        request=request,
-                        id=experiment_id,
-                    )
-
-            strat_id = self.configure(config=tempconfig)
-        else:
-            raise RuntimeError("Missing a configure message!")
-
-        return strat_id
-
     def handle_resume_v01(self, request):
         logger.debug("got resume message!")
         strat_id = int(request["message"]["strat_id"])
@@ -458,35 +255,17 @@ class AEPsychServer(object):
             )
         return self.strat_id
 
-    def handle_ask_v01(self, request):
-        """Returns dictionary with two entries:
-        "config" -- dictionary with config (keys are strings, values are floats)
-        "is_finished" -- bool, true if the strat is finished
-        """
-        logger.debug("got ask message!")
-        if self._pregen_asks:
-            params = self._pregen_asks.pop()
-        else:
-            params = self.ask()
-
-        new_config = {"config": params, "is_finished": self.strat.finished}
-        if not self.is_performing_replay:
-            self.db.record_message(
-                master_table=self._db_master_record, type="ask", request=request
-            )
-        return new_config
-
     def unversioned_handler(self, request):
         message_map = {
-            "setup": self.handle_setup,
-            "ask": self.handle_ask,
-            "tell": self.handle_tell,
+            "setup": handle_setup,
+            "ask": handle_ask,
+            "tell": handle_tell,
             "update": self.handle_update,
-            "query": self.handle_query,
+            "query": handle_query,
             "parameters": self.handle_params,
             "can_model": self.handle_can_model,
             "exit": self.handle_exit,
-            "get_config": self.handle_get_config,
+            "get_config": handle_get_config,
             "finish_strategy": self.handle_finish_strategy,
             "strategy_name": self.handle_strategy_name,
         }
@@ -497,7 +276,11 @@ class AEPsychServer(object):
             type = request["type"]
             if type in message_map.keys():
                 logger.info(f"Received msg [{type}]")
-                ret_val = message_map[type](request)
+                ret_val = (
+                    message_map[type](request)
+                    if type not in {"ask", "setup", "query", "get_config", "tell"}
+                    else message_map[type](self, request)
+                )
 
                 return ret_val
             else:
@@ -507,80 +290,13 @@ class AEPsychServer(object):
 
                 raise RuntimeError(exception_message)
 
-    def handle_setup(self, request):
-        logger.debug("got setup message!")
-
-        if not self.is_performing_replay:
-            experiment_id = None
-            if self._db_master_record is not None:
-                experiment_id = self._db_master_record.experiment_id
-
-            self._db_master_record = self.db.record_setup(
-                description=DEFAULT_DESC,
-                name=DEFAULT_NAME,
-                request=request,
-                id=experiment_id,
-            )
-
-        if (
-            "config_str" in request["message"].keys()
-            or "config_dict" in request["message"].keys()
-        ):
-
-            _ = self.configure(**request["message"])
-        else:
-            raise RuntimeError("Missing a configure message!")
-        new_config = self.handle_ask(request)
-
-        return new_config
-
-    def handle_ask(self, request):
-        logger.debug("got ask message!")
-
-        if self._pregen_asks:
-            return self._pregen_asks.pop()
-
-        new_config = self.ask()
-
-        if not self.is_performing_replay:
-            self.db.record_message(
-                master_table=self._db_master_record, type="ask", request=request
-            )
-
-        return new_config
-
-    def handle_tell(self, request):
-        logger.debug("got tell message!")
-
-        if not self.is_performing_replay:
-            self.db.record_message(
-                master_table=self._db_master_record, type="tell", request=request
-            )
-
-        # Batch update mode
-        if type(request["message"]) == list:
-            for msg in request["message"]:
-                self.tell(**msg)
-        else:
-            self.tell(**request["message"])
-
-        if self.strat is not None and self.strat.finished is True:
-            logger.info("Recording strat because the experiment is complete.")
-
-            buffer = io.BytesIO()
-            torch.save(self.strat, buffer, pickle_module=dill)
-            buffer.seek(0)
-            self.db.record_strat(master_table=self._db_master_record, strat=buffer)
-
-        return "acq"
-
     def handle_update(self, request):
         # update is syntactic sugar for tell, then ask
         logger.debug("got update message!")
 
-        self.handle_tell(request)
+        handle_tell(self, request)
 
-        new_config = self.handle_ask(request)
+        new_config = handle_ask(self, request)
 
         return new_config
 
@@ -606,67 +322,6 @@ class AEPsychServer(object):
             )
         return {"can_model": self.strat.can_fit}
 
-    def handle_query(self, request):
-        logger.debug("got query message!")
-        if not self.is_performing_replay:
-            self.db.record_message(
-                master_table=self._db_master_record, type="query", request=request
-            )
-        response = self.query(**request["message"])
-        return response
-
-    def query(
-        self,
-        query_type="max",
-        probability_space=False,
-        x=None,
-        y=None,
-        constraints=None,
-    ):
-        if self.skip_computations:
-            return None
-
-        constraints = constraints or {}
-        response = {
-            "query_type": query_type,
-            "probability_space": probability_space,
-            "constraints": constraints,
-        }
-        if query_type == "max":
-            fmax, fmax_loc = self.strat.get_max(constraints)
-            response["y"] = fmax.item()
-            response["x"] = self._tensor_to_config(fmax_loc)
-        elif query_type == "min":
-            fmin, fmin_loc = self.strat.get_min(constraints)
-            response["y"] = fmin.item()
-            response["x"] = self._tensor_to_config(fmin_loc)
-        elif query_type == "prediction":
-            # returns the model value at x
-            if x is None:  # TODO: ensure if x is between lb and ub
-                raise RuntimeError("Cannot query model at location = None!")
-            mean, var = self.strat.predict(
-                self._config_to_tensor(x).unsqueeze(axis=0),
-                probability_space=probability_space,
-            )
-            response["x"] = x
-            response["y"] = mean.item()
-        elif query_type == "inverse":
-            # expect constraints to be a dictionary; values are float arrays size 1 (exact) or 2 (upper/lower bnd)
-            constraints = {self.parnames.index(k): v for k, v in constraints.items()}
-            nearest_y, nearest_loc = self.strat.inv_query(
-                y, constraints, probability_space=probability_space
-            )
-            response["y"] = nearest_y
-            response["x"] = self._tensor_to_config(nearest_loc)
-        else:
-            raise RuntimeError("unknown query type!")
-        # ensure all x values are arrays
-        response["x"] = {
-            k: np.array([v]) if np.array(v).ndim == 0 else v
-            for k, v in response["x"].items()
-        }
-        return response
-
     def handle_exit(self, request):
         # Make local server write strats into DB and close the connection
         termination_type = "Normal termination"
@@ -677,24 +332,6 @@ class AEPsychServer(object):
 
         # If using thrift, it will add 'Terminate' to the queue and pass it to thrift server level
         return "Terminate"
-
-    def handle_get_config(self, request):
-        msg = request["message"]
-        section = msg.get("section", None)
-        prop = msg.get("property", None)
-
-        # If section and property are not specified, return the whole config
-        if section is None and prop is None:
-            return self.config.to_dict(deduplicate=False)
-
-        # If section and property are not both specified, raise an error
-        if section is None and prop is not None:
-            raise RuntimeError("Message contains a property but not a section!")
-        if section is not None and prop is None:
-            raise RuntimeError("Message contains a section but not a property!")
-
-        # If both section and property are specified, return only the relevant value from the config
-        return self.config.to_dict(deduplicate=False)[section][prop]
 
     def handle_finish_strategy(self, request):
         self.strat.finish()
@@ -745,27 +382,6 @@ class AEPsychServer(object):
     def can_pregen_ask(self):
         return self.strat is not None and self.enable_pregen
 
-    def ask(self):
-        """get the next point to query from the model
-        Returns:
-            dict -- new config dict (keys are strings, values are floats)
-        """
-        if self.skip_computations:
-            # HACK to makke sure strategies finish correctly
-            self.strat._strat._count += 1
-            if self.strat._strat.finished:
-                self.strat._make_next_strat()
-            return None
-
-        if not self.use_ax:
-            # index by [0] is temporary HACK while serverside
-            # doesn't handle batched ask
-            next_x = self.strat.gen()[0]
-            return self._tensor_to_config(next_x)
-
-        next_x = self.strat.gen()
-        return next_x
-
     def _tensor_to_config(self, next_x):
         config = {}
         for name, val in zip(self.parnames, next_x):
@@ -784,145 +400,6 @@ class AEPsychServer(object):
         else:
             x = torch.tensor(np.stack(unpacked))
         return x
-
-    def tell(self, outcome, config, model_data=True):
-        """tell the model which input was run and what the outcome was
-        Arguments:
-            inputs {dict} -- dictionary, keys are strings, values are floats or int.
-            keys should inclde all of the parameters we are tuning over, plus 'outcome'
-            which would be in {0, 1}.
-            TODO better types
-        """
-        if not self.is_performing_replay:
-            self._db_raw_record = self.db.record_raw(
-                master_table=self._db_master_record,
-                model_data=bool(model_data),
-            )
-
-            for param_name, param_value in config.items():
-                if isinstance(param_value, Iterable) and type(param_value) != str:
-                    if len(param_value) == 1:
-                        self.db.record_param(
-                            raw_table=self._db_raw_record,
-                            param_name=str(param_name),
-                            param_value=str(param_value[0]),
-                        )
-                    else:
-                        for i, v in enumerate(param_value):
-                            self.db.record_param(
-                                raw_table=self._db_raw_record,
-                                param_name=str(param_name) + "_stimuli" + str(i),
-                                param_value=str(v),
-                            )
-                else:
-                    self.db.record_param(
-                        raw_table=self._db_raw_record,
-                        param_name=str(param_name),
-                        param_value=str(param_value),
-                    )
-
-            if isinstance(outcome, dict):
-                for key in outcome.keys():
-                    self.db.record_outcome(
-                        raw_table=self._db_raw_record,
-                        outcome_name=key,
-                        outcome_value=float(outcome[key]),
-                    )
-
-            # Check if we get single or multiple outcomes
-            # Multiple outcomes come in the form of iterables that aren't strings or single-element tensors
-            elif isinstance(outcome, Iterable) and type(outcome) != str:
-                for i, outcome_value in enumerate(outcome):
-                    if (
-                        isinstance(outcome_value, Iterable)
-                        and type(outcome_value) != str
-                    ):
-                        if (
-                            isinstance(outcome_value, torch.Tensor)
-                            and outcome_value.dim() < 2
-                        ):
-                            outcome_value = outcome_value.item()
-
-                        elif len(outcome_value) == 1:
-                            outcome_value = outcome_value[0]
-                        else:
-                            raise ValueError(
-                                "Multi-outcome values must be a list of lists of length 1!"
-                            )
-                    self.db.record_outcome(
-                        raw_table=self._db_raw_record,
-                        outcome_name="outcome_" + str(i),
-                        outcome_value=float(outcome_value),
-                    )
-            else:
-                self.db.record_outcome(
-                    raw_table=self._db_raw_record,
-                    outcome_name="outcome",
-                    outcome_value=float(outcome),
-                )
-
-        if model_data:
-            if not self.use_ax:
-                x = self._config_to_tensor(config)
-            else:
-                x = config
-            self.strat.add_data(x, outcome)
-
-    def _configure(self, config):
-        self._pregen_asks = (
-            []
-        )  # TODO: Allow each strategy to have its own stack of pre-generated asks
-
-        parnames = config._str_to_list(
-            config.get("common", "parnames"), element_type=str
-        )
-        self.parnames = parnames
-        self.config = config
-        self.use_ax = config.getboolean("common", "use_ax", fallback=False)
-        self.enable_pregen = config.getboolean("common", "pregen_asks", fallback=False)
-        if self.use_ax:
-            self.trial_index = -1
-            self.strat = AEPsychStrategy.from_config(config)
-            self.strat_id = self.n_strats - 1
-
-        else:
-            self.strat = SequentialStrategy.from_config(config)
-            self.strat_id = self.n_strats - 1  # 0-index strats
-
-        return self.strat_id
-
-    def configure(self, config=None, **config_args):
-        # To preserve backwards compatibility, config_args is still usable for unittests and old functions.
-        # But if config is specified, the server will use that rather than create a new config object.
-        if config is None:
-            usedconfig = Config(**config_args)
-        else:
-            usedconfig = config
-        if "experiment" in usedconfig:
-            logger.warning(
-                'The "experiment" section is being deprecated from configs. Please put everything in the "experiment" section in the "common" section instead.'
-            )
-
-            for i in usedconfig["experiment"]:
-                usedconfig["common"][i] = usedconfig["experiment"][i]
-            del usedconfig["experiment"]
-
-        version = usedconfig.version
-        if version < __version__:
-            try:
-                usedconfig.convert_to_latest()
-
-                self.db.perform_updates()
-                logger.warning(
-                    f"Config version {version} is less than AEPsych version {__version__}. The config was automatically modified to be compatible. Check the config table in the db to see the changes."
-                )
-            except RuntimeError:
-                logger.warning(
-                    f"Config version {version} is less than AEPsych version {__version__}, but couldn't automatically update the config! Trying to configure the server anyway..."
-                )
-
-        self.db.record_config(master_table=self._db_master_record, config=usedconfig)
-        return self._configure(usedconfig)
 
     def __getstate__(self):
         # nuke the socket since it's not pickleble
@@ -961,7 +438,7 @@ def startServerAndRun(
         if config_path is not None:
             with open(config_path) as f:
                 config_str = f.read()
-            server.configure(config_str=config_str)
+            configure(server, config_str=config_str)
 
         if socket is not None:
             if uuid_of_replay is not None:
@@ -974,7 +451,7 @@ def startServerAndRun(
                     "You have passed in a config path but this is a replay. If there's a config in the database it will be used instead of the passed in config path."
                 )
             server.replay(uuid_of_replay)
-    except (KeyboardInterrupt):
+    except KeyboardInterrupt:
         exception_type = "CTRL+C"
         dump_type = "dump"
         server.write_strats(exception_type)
