+++ website/static/files/minimal_python_example.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # coding: utf-8
 
-# In this example, we demonstrate how to run a 2-dimensional detection-threshold experiment.  
+# In this example, we demonstrate how to run a 2-dimensional detection-threshold experiment.
 # **Please be aware that this package is still under development and the API is subject to change. This example will be updated as needed.**
 
 # First we fix the random number generators so that our code is reproducible:
@@ -12,29 +12,32 @@
 import numpy as np
 import torch
 
+from aepsych.server.message_handlers.handle_ask import ask
+
 
 # Fix random seeds
 np.random.seed(0)
 torch.manual_seed(0)
 
 
-# To use AEPsych, you have to create an experiment configuration file. For this example, we will use the example configuration file found under [aepsych/configs/single_lse_example.ini](https://github.com/facebookresearch/aepsych/blob/main/configs/single_lse_example.ini).  
-# On each trial of the experiment we show a single stimulus controlled by 2 parameters, par1 and par2. On each trial, the participant indicates whether or not they detected the stimulus. The goal is to find the detection threshold. See the contents of the configuration file for further explanation.  
-# 
+# To use AEPsych, you have to create an experiment configuration file. For this example, we will use the example configuration file found under [aepsych/configs/single_lse_example.ini](https://github.com/facebookresearch/aepsych/blob/main/configs/single_lse_example.ini).
+# On each trial of the experiment we show a single stimulus controlled by 2 parameters, par1 and par2. On each trial, the participant indicates whether or not they detected the stimulus. The goal is to find the detection threshold. See the contents of the configuration file for further explanation.
+#
 # The configuration file is used to configure the server object that will tell us which parameters to try on every trial:
 
 # In[ ]:
 
 
 from aepsych.server import AEPsychServer
+from aepsych.server.message_handlers.handle_setup import configure
 
 
 # Create a server object configured to run a 2d threshold experiment
 server = AEPsychServer()
-server.configure(config_fnames=['../configs/single_lse_example.ini'])
+configure(server, config_fnames=["../configs/single_lse_example.ini"])
 
 
-# In a real experiment you would write your own code to display stimuli and collect responses, but in this toy example we will simulate participant responses. We define the true 75% detection threshold to be where par1 + par2 = 1. We simulate participant responses using bernoulli trials, with 1 indicating a detection and 0 indicating no detection. 
+# In a real experiment you would write your own code to display stimuli and collect responses, but in this toy example we will simulate participant responses. We define the true 75% detection threshold to be where par1 + par2 = 1. We simulate participant responses using bernoulli trials, with 1 indicating a detection and 0 indicating no detection.
 
 # In[ ]:
 
@@ -47,15 +50,15 @@ from scipy.stats import norm
 # Define the 75% to be where par1 + par2 = 1
 def get_response_probability(params):
     m = 10
-    b = logit(.75) - m
+    b = logit(0.75) - m
     p = expit(m * params.sum(1) + b)
     return p
 
 
 # Simulate participant responses; returns 1 if the participant detected the stimulus or 0 if they did not.
 def simulate_response(trial_params):
-    par1 = trial_params['par1'][0]
-    par2 = trial_params['par2'][0]
+    par1 = trial_params["par1"][0]
+    par2 = trial_params["par2"][0]
     params = np.array([[par1, par2]])
 
     p = get_response_probability(params)
@@ -72,7 +75,7 @@ def simulate_response(trial_params):
 
 while not server.strat.finished:
     # Ask the server what the next parameter values to test should be.
-    trial_params = server.ask()
+    trial_params = ask(server)
 
     # Simulate a participant response.
     outcome = simulate_response(trial_params)
@@ -91,7 +94,7 @@ from scipy.stats import norm
 
 
 # Plot the results
-plot_strat(server.strat, title='Strategy Plot', true_testfun=get_response_probability)
+plot_strat(server.strat, title="Strategy Plot", true_testfun=get_response_probability)
 
 
 # ![Example plot](example_plot_strat_2d.png)
@@ -115,15 +118,15 @@ from scipy.stats import norm
 # Define the 75% to be where par1 + par2 = 1
 def get_response_probability(params):
     m = 10
-    b = logit(.75) - m
+    b = logit(0.75) - m
     p = expit(m * params.sum(1) + b)
     return p
 
 
 # Simulate participant responses; returns 1 if the participant detected the stimulus or 0 if they did not.
 def simulate_response(trial_params):
-    par1 = trial_params['par1'][0]
-    par2 = trial_params['par2'][0]
+    par1 = trial_params["par1"][0]
+    par2 = trial_params["par2"][0]
     params = np.array([[par1, par2]])
 
     p = get_response_probability(params)
@@ -138,7 +141,7 @@ torch.manual_seed(0)
 
 # Create a server object configured to run a 2d threshold experiment
 server = AEPsychServer()
-server.configure(config_fnames=['../configs/single_lse_example.ini'])
+configure(server, config_fnames=["../configs/single_lse_example.ini"])
 
 
 while not server.strat.finished:
@@ -152,5 +155,4 @@ while not server.strat.finished:
     server.tell(outcome, trial_params)
 
 # Plot the results
-plot_strat(server.strat, title='Strategy Plot', true_testfun=get_response_probability)
-
+plot_strat(server.strat, title="Strategy Plot", true_testfun=get_response_probability)
